requirements.txt
alembic==1.12.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
db-service-grpc @ git+https://github.com/alexseipopov/capybaras_db_handler_grpc.git@main
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
grpcio==1.59.2
grpcio-tools==1.59.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
protobuf==4.25.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
